tests/wpt/web-platform-tests/app-history/navigate/navigate-transitionWhile-history-state.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async () => {
  history.replaceState("state1", "", "#1");
  assert_equals(history.state, "state1");

  appHistory.onnavigate = e => e.transitionWhile(Promise.resolve("r"));
  let result = await appHistory.navigate("#2");
  assert_equals(result, undefined);
  assert_equals(location.hash, "#2");
  assert_equals(history.state, null);
}, "history.story should be nulled by navigate() handled by transitionWhile()");
</script>
